move non-task function into utils.invoicing
