+++ corehq/apps/accounting/tasks.py
@@ -37,12 +37,10 @@ from corehq.apps.accounting.models import (
     BillingAccount,
     CreditLine,
     Currency,
-    CustomerInvoice,
     DefaultProductPlan,
     DomainUserHistory,
     EntryPoint,
     FeatureType,
-    Invoice,
     InvoicingPlan,
     SoftwarePlanEdition,
     StripePaymentMethod,
@@ -66,7 +64,7 @@ from corehq.apps.accounting.utils import (
 )
 from corehq.apps.accounting.utils.invoicing import (
     get_domains_with_subscription_invoices_over_threshold,
-    UNPAID_INVOICE_THRESHOLD,
+    get_accounts_with_customer_invoices_over_threshold,
 )
 from corehq.apps.app_manager.dbaccessors import get_all_apps
 from corehq.apps.domain.models import Domain
@@ -746,35 +744,6 @@ def run_downgrade_process():
         subscription_on_invoice = oldest_unpaid_invoice.subscriptions.first()
         if is_subscription_eligible_for_downgrade_process(subscription_on_invoice):
             _apply_downgrade_process(oldest_unpaid_invoice, total, today)
-
-
-def get_accounts_with_customer_invoices_over_threshold(today):
-    unpaid_customer_invoices = CustomerInvoice.objects.filter(
-        is_hidden=False,
-        date_paid__isnull=True
-    )
-
-    overdue_customer_invoices_in_downgrade_daterange = unpaid_customer_invoices.filter(
-        date_due__lte=today - datetime.timedelta(days=1),
-        date_due__gte=today - datetime.timedelta(days=61)
-    ).order_by('date_due').select_related('account')
-
-    accounts = set()
-    for overdue_invoice in overdue_customer_invoices_in_downgrade_daterange:
-        account = overdue_invoice.account.name
-        plan = overdue_invoice.subscriptions.first().plan_version
-        if (account, plan) not in accounts:
-            invoices = unpaid_customer_invoices.filter(
-                Q(date_due__lte=overdue_invoice.date_due)
-                | (Q(date_due__isnull=True) & Q(date_end__lte=overdue_invoice.date_end)),
-                account__name=account
-            )
-            invoices = [invoice for invoice in invoices if invoice.subscriptions.first().plan_version == plan]
-            total_overdue_to_date = sum(invoice.balance for invoice in invoices)
-
-            if total_overdue_to_date >= UNPAID_INVOICE_THRESHOLD:
-                accounts.add((account, plan))
-                yield overdue_invoice, total_overdue_to_date
 
 
 def is_subscription_eligible_for_downgrade_process(subscription):
